+++ src/main/routes/forget-password-routes.ts
@@ -1,6 +1,6 @@
 import { type Router } from 'express'
-import { adaptRoute } from '@/main/adapters/express-route-adapter'
-import { makeForgetPasswordController } from '../factories/controllers/forget-password-factory'
+import { adaptRoute } from '@/main/adapters'
+import { makeForgetPasswordController } from '@/main/factories'
 
 export default (router: Router): void => {
   router.post('/forget-password', adaptRoute(makeForgetPasswordController()))
