+++ src/main/routes/login-routes.ts
@@ -1,6 +1,6 @@
 import { type Router } from 'express'
-import { adaptRoute } from '@/main/adapters/express-route-adapter'
-import { makeLoginController } from '@/main/factories/controllers/login-controller-factory'
+import { adaptRoute } from '@/main/adapters'
+import { makeLoginController } from '@/main/factories'
 
 export default (router: Router): void => {
   router.post('/login', adaptRoute(makeLoginController()))
