+++ src/main/routes/signup-routes.ts
@@ -1,6 +1,6 @@
 import { type Router } from 'express'
-import { adaptRoute } from '@/main/adapters/express-route-adapter'
-import { makeSignUpController } from '@/main/factories/controllers/signup-controller-factory'
+import { adaptRoute } from '@/main/adapters'
+import { makeSignUpController } from '@/main/factories'
 
 export default (router: Router): void => {
   router.post('/signup', adaptRoute(makeSignUpController()))
